feat: add import command and update product pricing structure
- Introduced a new command to import products from Elasticsearch, enhancing data management capabilities.
- Updated the products table schema to include price and origin_price fields, allowing for better pricing management.
- Changed price fields in product-related structures from float64 to int64 for consistency and improved performance.
- Adjusted product filtering and handling logic to accommodate the new pricing structure, ensuring accurate data processing.

+++ internal/products/domain/entity/product_aggregate/product_detail.go
@@ -25,8 +25,8 @@ type ProductDetail struct {
 	ColorKey      *string             `json:"color_key,omitempty" gorm:"color_key"`   // e.g., "red", "blue" - nullable for size-only details
 	ColorName     *string             `json:"color_name,omitempty" gorm:"color_name"` // e.g., "قرمز", "آبی" - nullable
 	SizeKey       *string             `json:"size_key,omitempty" gorm:"size_key"`     // e.g., "M", "L" - nullable for color-only details
-	Price         float64             `json:"price" gorm:"price"`                     // Price for this variant
-	OriginalPrice *float64            `json:"-" gorm:"original_price"`                // Original price before discount (deprecated, not exposed in API)
+	Price         int64               `json:"price" gorm:"price"`                     // Price for this variant
+	OriginalPrice *int64              `json:"-" gorm:"original_price"`                // Original price before discount (deprecated, not exposed in API)
 	Stock         int                 `json:"stock" gorm:"stock;default:0"`
 	Discount      int                 `json:"discount" gorm:"discount;default:0"`
 	Images        []shared.Attachment `json:"-" gorm:"polymorphic:Attachable;polymorphicValue:ProductDetail"` // Polymorphic relationship
